Util: Use write-sync-close-rename for config file writing

+++ core/java/src/net/i2p/util/FileUtil.java
@@ -480,11 +480,12 @@ public class FileUtil {
         boolean success = false;
         boolean isWindows = SystemVersion.isWindows();
         // overwrite fails on windows
-        if (!isWindows)
+        boolean exists = to.exists();
+        if (!isWindows || !exists)
             success = from.renameTo(to);
         if (!success) {
-            to.delete();
-            success = from.renameTo(to);
+            if (exists && to.delete())
+                success = from.renameTo(to);
             if (!success) {
                 // hard way
                 success = copy(from, to, true, true);
@@ -496,12 +497,12 @@ public class FileUtil {
     }
 
     /**
-     * Usage: FileUtil (delete path | copy source dest | unzip path.zip)
+     * Usage: FileUtil (delete path | copy source dest | rename from to | unzip path.zip)
      *
      */
     public static void main(String args[]) {
         if ( (args == null) || (args.length < 2) ) {
-            System.err.println("Usage: delete path | copy source dest | unzip path.zip");
+            System.err.println("Usage: delete path | copy source dest | rename from to | unzip path.zip");
             //testRmdir();
         } else if ("delete".equals(args[0])) {
             boolean deleted = FileUtil.rmdir(args[1], false);
@@ -523,6 +524,12 @@ public class FileUtil {
                 System.err.println("Unzipped [" + args[1] + "] to [" + to + "]");
             else
                 System.err.println("Error unzipping [" + args[1] + "] to [" + to + "]");
+        } else if ("rename".equals(args[0])) {
+            boolean success = rename(new File(args[1]), new File(args[2]));
+            if (!success) 
+                System.err.println("Error renaming [" + args[1] + "] to [" + args[2] + "]");
+        } else {
+            System.err.println("Usage: delete path | copy source dest | rename from to | unzip path.zip");
         }
     }
